Moved tools execution into agent tool registry and agent tool handlers
